app/requirements.txt
flask
pymongo
flask-mongoengine
Flask-WTF
flask-login
flask-restful
Flask-JWT
surprise
pandas
googletrans==3.1.0a0
scikit-surprise
scikit-learn
google-api-python-client
oauth2client
Flask-Caching
xmltodict
Sip2
Flask-APScheduler
line-bot-sdk
pythainlp==2.3.2
spacy_thai
deplacy==2.0.2
fairseq==0.10.0
sacremoses==0.0.47
sentencepiece==0.1.96
antlr4-python3-runtime==4.8
blis==0.7.5
cachetools==4.2.4
catalogue==2.0.6
certifi==2021.10.8
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.10
click==8.0.3
colorama==0.4.4
cymem==2.0.6
Cython==0.29.26
dataclasses==0.6
editdistance==0.6.0
en-core-web-md @ https://github.com/explosion/spacy-models/releases/download/en_core_web_md-3.2.0/en_core_web_md-3.2.0-py3-none-any.whl
google-api-core==1.31.5
google-auth==1.35.0
google-cloud-core==1.7.2
google-cloud-translate==2.0.1
googleapis-common-protos==1.54.0
grpcio==1.43.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2021.12.1
httpcore==0.9.1
httpx==0.13.3
hydra-core==1.1.1
hyperframe==5.2.0
idna==2.10
importlib-resources==5.4.0
Jinja2==3.0.3
joblib==1.1.0
langcodes==3.3.0
MarkupSafe==2.0.1
murmurhash==1.0.6
nltk==3.6.7
numpy==1.22.0
omegaconf==2.1.1
packaging==21.3
pathy==0.6.1
plac==1.1.3
portalocker==2.3.2
preshed==3.0.6
protobuf==3.19.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.8.2
pyparsing==3.0.6
python-crfsuite==0.9.7
pytz==2021.3
PyYAML==6.0
regex==2021.11.10
requests==2.27.1
rfc3986==1.5.0
rsa==4.8
sacrebleu==2.0.0
six==1.16.0
smart-open==5.2.1
sniffio==1.2.0
spacy==3.2.1
spacy-legacy==3.0.8
spacy-loggers==1.0.1
spacy-thai==0.6.7
spacytextblob==3.0.1
srsly==2.4.2
tabulate==0.8.9
textblob==0.15.3
thinc==8.0.13
tinydb==4.5.2
torch==1.10.1
tqdm==4.62.3
typer==0.4.0
typing-extensions==4.0.1
ufal.udpipe==1.2.0.3
urllib3==1.26.8
wasabi==0.9.0
zipp==3.7.0
faker
Flask-Session
